+++ assert_http.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/beme/abide/utils"
+	"github.com/beme/abide/internal"
 	"github.com/nsf/jsondiff"
 	"github.com/sergi/go-diff/diffmatchpatch"
 )
@@ -70,7 +70,7 @@ func (h *httpRequest) jsonBodyCleanup(c *config) error {
 	// Clean/update json based on config.
 	if c != nil {
 		for k, v := range c.Defaults {
-			jsonIface = utils.UpdateKeyValuesInMap(k, v, jsonIface)
+			jsonIface = internal.UpdateKeyValuesInMap(k, v, jsonIface)
 		}
 	}
